+++ pkg/backends/logger.go
@@ -1,0 +1,9 @@
+package backends
+
+import (
+	zlog "github.com/liut/staffio/pkg/log"
+)
+
+func logger() zlog.Logger {
+	return zlog.GetLogger()
+}
