+++ pkg/backends/wechatwork/logger.go
@@ -1,0 +1,9 @@
+package wechatwork
+
+import (
+	zlog "github.com/liut/staffio/pkg/log"
+)
+
+func logger() zlog.Logger {
+	return zlog.GetLogger()
+}
